+++ sdk/storage/Azure.Storage.DataMovement/src/Models/JobPlan/JobPartPlanDestinationLocal.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-namespace Azure.Storage.DataMovement.JobPlanModels
+namespace Azure.Storage.DataMovement.Models.JobPlan
 {
     /// <summary>
     /// This matching the JobPartPlanDstLocal of azcopy
