+++ sdk/storage/Azure.Storage.DataMovement/tests/JobPartPlanFileTests.cs
@@ -4,8 +4,8 @@
 using System.IO;
 using System.Threading.Tasks;
 using NUnit.Framework;
-using Azure.Storage.DataMovement.JobPlanModels;
 using Azure.Core.TestFramework;
+using Azure.Storage.DataMovement.Models.JobPlan;
 
 namespace Azure.Storage.DataMovement.Tests
 {
