+++ sdk/storage/Azure.Storage.DataMovement/tests/LocalTransferCheckpointerTests.cs
@@ -9,8 +9,8 @@ using System.Text;
 using System.Threading.Tasks;
 using NUnit.Framework;
 using Azure.Storage.DataMovement.Models;
-using Azure.Storage.DataMovement.JobPlanModels;
 using Azure.Core.TestFramework;
+using Azure.Storage.DataMovement.Models.JobPlan;
 
 namespace Azure.Storage.DataMovement.Tests
 {
