+++ sdk/storage/Azure.Storage.DataMovement/tests/Shared/DataMovementBlobTestBase.cs
@@ -18,9 +18,6 @@ using Azure.Core;
 using System.Threading;
 using Azure.Storage.Blobs.Tests;
 using Azure.Storage.Shared;
-using Azure.Storage.DataMovement.JobPlanModels;
-using System.Drawing;
-using Azure.Storage.DataMovement.Models;
 
 namespace Azure.Storage.DataMovement.Tests
 {
